Mapping the π key to 'π' in eqn editor in STD mode
This is just cosmetic, since both 'π' and 'PI' are recognized as the
constant π, in both STD and COMP modes. I'm just keeping the π -> 'PI'
mapping in COMP mode because the HP calculators that COMP mode is
supposed to be compatible with don't have the π character, so using 'PI'
helps with two-way compatibility.

Also, added π to the ALPHA menu, under MATH.