Turning off optimization for MacOS
Test case: BRESET 64 WSIZE 1 ENTER BASE÷
Should return 1, does return Out of Range

The culprit is the test in base_range_check(), in core_helpers.cc line
959 (Free42: line 810), which is getting sabotaged by an optimizer bug
in the iOS and Android release builds. I'm turning optimization off for
MacOS as well, since the problem is in the clang ARM back-end; given
that this affects the Android and iOS release builds, and not the
x64-based simulator builds, nor the Windows and Linux builds, it seems
likely that the ARM code for MacOS is affected in the same way, even if
I don't have an M1 Mac handy to test this right now. (The DM42 seems
immune, its toolchain may use a different compiler or maybe SM just
plays safe and builds their firmware with -O0 anyway.)