PLOT/SOLVE message fix
The existing logic did make sense after all, I just didn't get it when I
committed the PLOT/SOLVE fix. I just needed to add the check for state
== PLOT_STATE_SOLVE to the check for state != PLOT_STATE_INTEG &&
data.axes[1].len > 0, not replace the latter with the former. The latter
is the check for SOLVE being used for the main plot or for SCAN, which
is supposed to be handled like a function that returns an error, i.e.
not plot a point, and because of my mistake, this is now handled like
the function returning 0. Since 0 is where the X axis is, the difference
will usually not even be noticeable, but it does make a difference when
there are discontinuities, and of course it's just ugly, so here's the
proper fix. Hopefully the last one!