Fixed Intel lib tests on arm64
The tests now run cleanly on Linux. On MacOS, there are still errors in
the conversions to and from binary80, caused by the lack of long double
support in Apple's compiler. Not a big deal since Plus42 doesn't use
those functions.

Still to be done: Windows on ARM, but of course, that's a bit of a
project in and of itself, since I don't even have a Windows ARM build
yet, so who knows what happens with the Intel library in that scenario.
This should be dealt with eventually, but there's no rush.