Removed unnecessary comment
The comment was about CodeMap::lookup(), that it might need to go up the
RTN stack if it can't resolve the current pc. Thinking about it some
more, I can't even come up with any scenario when that might happen;
remember that CodeMap::lookup() is only called when the program
directory is 1.
Should I ever do hear about, or run into, a scenario where
CodeMap::lookup() does need to retry with caller pc's, I'll deal with it
then. I could do it now (the modification itself is simple) but there is
no point if I don't even know how to test it.