Fixed err msg for non-writable var with STO arith
It used to report "Nonexistent" because it wouldn't even look outside of
the current directory. That was a bad performance optimization: in these
cases, if the variable isn't found in the current directory, it's an
error anyway, and so the added cost of distinguishing between
"Nonexistent" and "Variable Not Writable" is irrelevant, while the
difference between those errors to the user can be very relevant.