Using Unicode ratio sign for thin colon
Just so we can copy and paste it without turning it into a regular colon. It
will probably not impact much, since this is a character that never seems to be
used outside of labeling the stack registers in the RUN mode display. And
having a distinct representation for it makes it easier to support it in the
alpha keyboard.